+++ src/assets/mod.rs
@@ -1,10 +1,8 @@
+mod asset;
 mod asset_loader;
 mod parser;
 
-pub(crate) mod vector;
-
-pub use asset_loader::VelloAssetLoader;
-pub use vector::{VectorFile, VelloAsset};
-
+pub use asset::{VectorFile, VelloAsset};
 pub(crate) use asset_loader::VectorLoaderError;
+pub(crate) use asset_loader::VelloAssetLoader;
 pub use parser::*;