+++ src/text/font.rs
@@ -14,13 +14,8 @@
 //
 // Also licensed under MIT license, at your choice.
 
-use bevy::{
-    asset::{io::Reader, AssetLoader, AsyncReadExt, LoadContext},
-    prelude::*,
-    reflect::TypePath,
-    render::render_asset::RenderAsset,
-    utils::BoxedFuture,
-};
+use super::vello_text::VelloText;
+use bevy::{prelude::*, reflect::TypePath, render::render_asset::RenderAsset};
 use std::sync::Arc;
 use vello::{
     fello::{meta::MetadataProvider, raw::FontRef},
@@ -29,8 +24,6 @@ use vello::{
     peniko::{self, Blob, Brush, Font},
     SceneBuilder, SceneFragment,
 };
-
-use crate::assets::VectorLoaderError;
 
 #[derive(Asset, TypePath)]
 pub struct VelloFont {
@@ -71,34 +64,63 @@ impl VelloFont {
         }
     }
 
-    pub fn render_centered(
+    pub fn sizeof(&self, text: &VelloText) -> Vec2 {
+        let font = FontRef::new(self.font.data.data())
+            .expect("Vello font creation error");
+
+        let fello_size = vello::fello::Size::new(text.size);
+        let charmap = font.charmap();
+        let metrics = font.metrics(fello_size, Default::default());
+        let line_height = metrics.ascent - metrics.descent + metrics.leading;
+        let glyph_metrics = font.glyph_metrics(fello_size, Default::default());
+
+        let mut pen_x = 0.0;
+        let mut pen_y: f32 = 0.0;
+        let mut width: f32 = 0.0;
+        let mut height: f32 = line_height;
+        for ch in text.content.chars() {
+            if ch == '\n' {
+                pen_y += line_height;
+                pen_x = 0.0;
+                continue;
+            }
+            let gid = charmap.map(ch).unwrap_or_default();
+            let advance = glyph_metrics.advance_width(gid).unwrap_or_default();
+            pen_x += advance;
+            width = width.max(pen_x);
+        }
+        height += pen_y;
+        Vec2::new(width, height)
+    }
+
+    pub fn render(
         &mut self,
         builder: &mut SceneBuilder,
-        size: f32,
         transform: Affine,
-        text: &str,
+        text: &VelloText,
     ) {
-        let (glyphs, text_width, _text_height) =
-            self.add(size, None, transform, text);
-        for (glyph, xform) in glyphs {
-            let xform = xform * Affine::translate((-text_width / 2.0, 0.0));
-            builder.append(&glyph, Some(xform));
+        let frags = self.render_fragments(
+            text.size,
+            text.brush.as_ref(),
+            transform,
+            &text.content,
+        );
+        for (glyph, transform) in frags {
+            builder.append(&glyph, Some(transform));
         }
     }
 
-    fn add(
+    pub(crate) fn render_fragments(
         &mut self,
         size: f32,
         brush: Option<&Brush>,
         transform: Affine,
         text: &str,
-    ) -> (Vec<(SceneFragment, Affine)>, f64, f64) {
+    ) -> Vec<(SceneFragment, Affine)> {
         let font = FontRef::new(self.font.data.data())
             .expect("Vello font creation error");
 
         let mut items = vec![];
-        let mut pen_x = 0f64;
-        let mut pen_y = 0f64;
 
         let fello_size = vello::fello::Size::new(size);
         let charmap = font.charmap();
@@ -108,6 +130,9 @@ impl VelloFont {
         let vars: [(&str, f32); 0] = [];
         let mut provider =
             self.gcx.new_provider(&font, None, size, false, vars);
+
+        let mut pen_x: f64 = 0.0;
+        let mut pen_y: f64 = 0.0;
 
         for ch in text.chars() {
             if ch == '\n' {
@@ -121,42 +146,17 @@ impl VelloFont {
 
             if let Some(glyph) = provider.get(gid.to_u16(), brush) {
                 let xform = transform
-                    * Affine::translate((pen_x, 0.0))
+                    * Affine::translate((pen_x, pen_y))
                     * Affine::scale_non_uniform(1.0, -1.0);
                 // builder.append(&glyph, Some(xform));
                 items.push((glyph, xform));
             }
             pen_x += advance;
         }
-        (items, pen_x, pen_y)
-    }
-}
-#[derive(Default)]
-pub struct VelloFontLoader;
-
-impl AssetLoader for VelloFontLoader {
-    type Asset = VelloFont;
-
-    type Settings = ();
-
-    type Error = VectorLoaderError;
-
-    fn load<'a>(
-        &'a self,
-        reader: &'a mut Reader,
-        _settings: &'a Self::Settings,
-        _load_context: &'a mut LoadContext,
-    ) -> BoxedFuture<'a, Result<Self::Asset, Self::Error>> {
-        Box::pin(async move {
-            let mut bytes = Vec::new();
-            reader.read_to_end(&mut bytes).await?;
-            let vello_font = VelloFont::new(bytes.to_vec());
-
-            Ok(vello_font)
-        })
-    }
-
-    fn extensions(&self) -> &[&str] {
-        &["vttf"]
+        items
+            .into_iter()
+            // Push all lines up to account for new lines
+            .map(|(f, a)| (f, a * Affine::translate((0.0, pen_y))))
+            .collect()
     }
 }