+++ src/text/mod.rs
@@ -1,0 +1,7 @@
+mod font;
+mod font_loader;
+mod vello_text;
+
+pub use font::VelloFont;
+pub(crate) use font_loader::VelloFontLoader;
+pub use vello_text::VelloText;